+++ oxe-api/test/resource/private/test_add_my_form_answer.py
@@ -44,3 +44,4 @@ class TestAddFormAnswer(BaseCase):
         self.assertEqual(200, response.status_code)
         self.assertEqual(len(answers), 1)
         self.assertEqual(answers[0].value, "Text2")
+        self.assertEqual(answers[0].last_date, answers[0].sys_date)